services/analytics-engine/requirements.txt
flask==3.0.0
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0